fw: move @fw_signaling_cmd_handler into sub-modules
